moved TestInitialization to a new package (testingutils) and updated its dependent packages
