+++ src/test/java/io/jenkins/plugins/scmfilter/traitimplementations/SourceRegexSCMPrFilterTraitTest.java
@@ -2,8 +2,8 @@ package io.jenkins.plugins.scmfilter.traitimplementations;
 
 import static org.junit.Assert.assertTrue;
 
-import io.jenkins.plugins.scmfilter.TestInitialization;
 import io.jenkins.plugins.scmfilter.abstractclasses.SCMPrFilterTrait;
+import io.jenkins.plugins.scmfilter.testingutils.TestInitialization;
 import java.util.List;
 import jenkins.scm.api.trait.SCMHeadPrefilter;
 import jenkins.scm.api.trait.SCMSourceContext;
